# Use an on-stack value rather than a class