```
Benchmark                                                                              Mode  Cnt     Score      Error   Units
SlsVersionMatcherBenchmark.fast_matcher_with_all_numbers                               avgt    3    30.526 ±   17.676   ns/op
SlsVersionMatcherBenchmark.fast_matcher_with_all_numbers:·gc.alloc.rate                avgt    3  5998.483 ± 3474.348  MB/sec
SlsVersionMatcherBenchmark.fast_matcher_with_all_numbers:·gc.alloc.rate.norm           avgt    3    56.000 ±    0.001    B/op
SlsVersionMatcherBenchmark.fast_matcher_with_all_numbers:·gc.churn.G1_Eden_Space       avgt    3  6027.266 ± 7878.524  MB/sec
SlsVersionMatcherBenchmark.fast_matcher_with_all_numbers:·gc.churn.G1_Eden_Space.norm  avgt    3    56.221 ±   41.120    B/op
SlsVersionMatcherBenchmark.fast_matcher_with_all_numbers:·gc.churn.G1_Old_Gen          avgt    3     0.010 ±    0.079  MB/sec
SlsVersionMatcherBenchmark.fast_matcher_with_all_numbers:·gc.churn.G1_Old_Gen.norm     avgt    3    ≈ 10⁻⁴               B/op
SlsVersionMatcherBenchmark.fast_matcher_with_all_numbers:·gc.count                     avgt    3    42.000             counts
SlsVersionMatcherBenchmark.fast_matcher_with_all_numbers:·gc.time                      avgt    3    41.000                 ms
SlsVersionMatcherBenchmark.fast_matcher_with_one_x_x                                   avgt    3    21.459 ±    3.705   ns/op
SlsVersionMatcherBenchmark.fast_matcher_with_one_x_x:·gc.alloc.rate                    avgt    3  8528.711 ± 1517.903  MB/sec
SlsVersionMatcherBenchmark.fast_matcher_with_one_x_x:·gc.alloc.rate.norm               avgt    3    56.000 ±    0.001    B/op
SlsVersionMatcherBenchmark.fast_matcher_with_one_x_x:·gc.churn.G1_Eden_Space           avgt    3  8614.823 ± 5939.139  MB/sec
SlsVersionMatcherBenchmark.fast_matcher_with_one_x_x:·gc.churn.G1_Eden_Space.norm      avgt    3    56.556 ±   29.245    B/op
SlsVersionMatcherBenchmark.fast_matcher_with_one_x_x:·gc.churn.G1_Old_Gen              avgt    3     0.008 ±    0.022  MB/sec
SlsVersionMatcherBenchmark.fast_matcher_with_one_x_x:·gc.churn.G1_Old_Gen.norm         avgt    3    ≈ 10⁻⁴               B/op
SlsVersionMatcherBenchmark.fast_matcher_with_one_x_x:·gc.count                         avgt    3    46.000             counts
SlsVersionMatcherBenchmark.fast_matcher_with_one_x_x:·gc.time                          avgt    3    48.000                 ms
SlsVersionMatcherBenchmark.matcher_with_all_numbers                                    avgt    3   201.023 ±   15.323   ns/op
SlsVersionMatcherBenchmark.matcher_with_all_numbers:·gc.alloc.rate                     avgt    3  6892.536 ±  508.389  MB/sec
SlsVersionMatcherBenchmark.matcher_with_all_numbers:·gc.alloc.rate.norm                avgt    3   424.000 ±    0.006    B/op
SlsVersionMatcherBenchmark.matcher_with_all_numbers:·gc.churn.G1_Eden_Space            avgt    3  6851.703 ± 5390.440  MB/sec
SlsVersionMatcherBenchmark.matcher_with_all_numbers:·gc.churn.G1_Eden_Space.norm       avgt    3   421.446 ±  301.369    B/op
SlsVersionMatcherBenchmark.matcher_with_all_numbers:·gc.churn.G1_Old_Gen               avgt    3     0.007 ±    0.043  MB/sec
SlsVersionMatcherBenchmark.matcher_with_all_numbers:·gc.churn.G1_Old_Gen.norm          avgt    3    ≈ 10⁻³               B/op
SlsVersionMatcherBenchmark.matcher_with_all_numbers:·gc.count                          avgt    3    40.000             counts
SlsVersionMatcherBenchmark.matcher_with_all_numbers:·gc.time                           avgt    3    39.000                 ms
SlsVersionMatcherBenchmark.matcher_with_one_x_x                                        avgt    3   191.156 ±   11.748   ns/op
SlsVersionMatcherBenchmark.matcher_with_one_x_x:·gc.alloc.rate                         avgt    3  7658.757 ±  454.650  MB/sec
SlsVersionMatcherBenchmark.matcher_with_one_x_x:·gc.alloc.rate.norm                    avgt    3   448.000 ±    0.005    B/op
SlsVersionMatcherBenchmark.matcher_with_one_x_x:·gc.churn.G1_Eden_Space                avgt    3  7551.611 ± 6272.413  MB/sec
SlsVersionMatcherBenchmark.matcher_with_one_x_x:·gc.churn.G1_Eden_Space.norm           avgt    3   441.705 ±  349.671    B/op
SlsVersionMatcherBenchmark.matcher_with_one_x_x:·gc.churn.G1_Old_Gen                   avgt    3     0.005 ±    0.006  MB/sec
SlsVersionMatcherBenchmark.matcher_with_one_x_x:·gc.churn.G1_Old_Gen.norm              avgt    3    ≈ 10⁻⁴               B/op
SlsVersionMatcherBenchmark.matcher_with_one_x_x:·gc.count                              avgt    3    38.000             counts
SlsVersionMatcherBenchmark.matcher_with_one_x_x:·gc.time                               avgt    3    42.000                 ms
```

## Before this PR
<!-- What's wrong with the current state of the world and why change it now? -->

## After this PR
<!-- User-facing outcomes this PR delivers go below -->
==COMMIT_MSG==
==COMMIT_MSG==

## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->